python_version/requirements.txt
langchain>=0.0.300
chromadb>=0.4.15
sentence-transformers>=2.2.2
flask>=2.0.1
python-dotenv>=0.19.0
openai>=1.0.0
tiktoken>=0.5.1
# Add new dependencies for file handling
PyPDF2>=3.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-docx>=0.8.11
